# openapi.yaml
openapi: 3.0.3
info:
  title: parseable
  description: Parseable API documents [https://www.parseable.com/docs/](https://www.parseable.com/docs/)
  contact:
    name: Parseable
    url: https://www.parseable.com/
    email: hi@parseable.com
  license:
    name: ''
  version: 1.4.0
paths:
...
  /api/v1/user:
    get:
      tags:
      - User Management
      operationId: Get all users
      responses:
        '200':
          description: Fetches all users in the system
          content:
            application/json:
              schema:
                type: object
      security:
      - basic_auth: []
  /api/v1/user/{username}:
    post:
      tags:
      - User Management
      operationId: Create a user
      parameters:
      - name: username
        in: path
        description: username of the user to be created
        required: true
        schema:
          type: string
      requestBody:
        description: List of roles to be assigned to the user
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        '200':
          description: Created user and returned the initial password
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Error
        '404':
          description: User not found
        '500':
          description: Failure
      security:
      - basic_auth: []
    delete:
      tags:
      - User Management
      operationId: Delete a user
      parameters:
      - name: username
        in: path
        description: username of the user to be deleted
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Deleted the user
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Error
        '404':
          description: User not found
        '500':
          description: Failure
      security:
      - basic_auth: []
  /api/v1/user/{username}/generate-new-password:
    post:
      tags:
      - User Management
      operationId: Set a new password for user
      parameters:
      - name: username
        in: path
        description: username of the user
        required: true
        schema:
          type: string
      requestBody:
        description: Password for the user
        content:
          text/plain:
            schema:
              type: string
        required: true
      responses:
        '200':
          description: Assigned new password
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Error
        '404':
          description: User not found
        '500':
          description: Failure
      security:
      - basic_auth: []
  /api/v1/user/{username}/role:
    get:
      tags:
      - User Management
      operationId: Get a user's roles
      parameters:
      - name: username
        in: path
        description: username for which the role needs to be fetched
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Fetched role for the given user
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Error
        '404':
          description: User not found
        '500':
          description: Failure
      security:
      - basic_auth: []
    put:
      tags:
      - RBAC
      operationId: Add role to a user
      parameters:
      - name: username
        in: path
        description: username to which the roles need to be assigned
        required: true
        schema:
          type: string
      requestBody:
        description: Roles to be assigned to the user
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
        required: true
      responses:
        '200':
          description: Assigned roles to the user
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Error
        '404':
          description: User not found
        '500':
          description: Failure
      security:
      - basic_auth: []
components:
...
  securitySchemes:
    basic_auth:
      type: http
      scheme: basic
...